move components to own folders
