+++ docs/process-services/components/process-instance-tasks.component.md
@@ -5,7 +5,7 @@ Status: Active
 Last reviewed: 2018-11-19
 ---
 
-# [Process Instance Tasks component](../../../lib/process-services/src/lib/process-list/components/process-instance-tasks.component.ts "Defined in process-instance-tasks.component.ts")
+# [Process Instance Tasks component](../../../lib/process-services/src/lib/process-list/components/process-instance-tasks/process-instance-tasks.component.ts "Defined in process-instance-tasks.component.ts")
 
 Lists both the active and completed tasks associated with a particular process instance
