+++ lib/process-services/src/lib/process-list/components/process-filters/process-filters.component.spec.ts
@@ -17,12 +17,12 @@
 
 import { CUSTOM_ELEMENTS_SCHEMA, SimpleChange } from '@angular/core';
 import { from, of, throwError } from 'rxjs';
-import { AppsProcessService } from '../../app-list/services/apps-process.service';
-import { ProcessFilterService } from '../services/process-filter.service';
+import { AppsProcessService } from '../../../app-list/services/apps-process.service';
+import { ProcessFilterService } from '../../services/process-filter.service';
 import { ProcessFiltersComponent } from './process-filters.component';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { ProcessTestingModule } from '../../testing/process.testing.module';
+import { ProcessTestingModule } from '../../../testing/process.testing.module';
 import { NavigationStart, Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ProcessInstanceFilterRepresentation, UserProcessInstanceFilterRepresentation } from '@alfresco/js-api';
