+++ lib/process-services/src/lib/process-list/components/process-filters/process-filters.component.ts
@@ -18,9 +18,9 @@
 import { Component, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, SimpleChanges, ViewEncapsulation } from '@angular/core';
 import { ProcessInstanceFilterRepresentation, UserProcessInstanceFilterRepresentation } from '@alfresco/js-api';
 import { Subject } from 'rxjs';
-import { ProcessFilterService } from './../services/process-filter.service';
-import { AppsProcessService } from '../../app-list/services/apps-process.service';
-import { IconModel } from '../../app-list/icon.model';
+import { ProcessFilterService } from '../../services/process-filter.service';
+import { AppsProcessService } from '../../../app-list/services/apps-process.service';
+import { IconModel } from '../../../app-list/icon.model';
 import { NavigationStart, Router } from '@angular/router';
 import { filter, takeUntil } from 'rxjs/operators';
 import { CommonModule, Location } from '@angular/common';
