+++ lib/process-services/src/lib/process-list/components/process-instance-details/process-instance-details.component.css
