+++ lib/process-services/src/lib/process-list/components/process-instance-details/process-instance-details.component.spec.ts
@@ -19,14 +19,14 @@ import { NO_ERRORS_SCHEMA, SimpleChange } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { of } from 'rxjs';
-import { CommentProcessService } from '../../process-comments/services/comment-process.service';
-import { TaskListModule } from '../../task-list/task-list.module';
-import { exampleProcess, exampleProcessNoName, mockRunningProcess, processEnded } from './../../mock';
-import { mockProcessInstanceComments } from './../../mock/process/process-comments.mock';
-import { ProcessService } from './../services/process.service';
+import { CommentProcessService } from '../../../process-comments/services/comment-process.service';
+import { TaskListModule } from '../../../task-list/task-list.module';
+import { exampleProcess, exampleProcessNoName, mockRunningProcess, processEnded } from '../../../mock';
+import { mockProcessInstanceComments } from '../../../mock/process/process-comments.mock';
+import { ProcessService } from '../../services/process.service';
 import { ProcessInstanceDetailsComponent } from './process-instance-details.component';
-import { ProcessTestingModule } from '../../testing/process.testing.module';
-import { FormModule } from '../../form';
+import { ProcessTestingModule } from '../../../testing/process.testing.module';
+import { FormModule } from '../../../form';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { MatCardHarness } from '@angular/material/card/testing';
