+++ lib/process-services/src/lib/process-list/components/process-instance-header/process-instance-header.component.spec.ts
@@ -17,9 +17,9 @@
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { AppConfigService } from '@alfresco/adf-core';
-import { exampleProcess } from '../../mock';
+import { exampleProcess } from '../../../mock';
 import { ProcessInstanceHeaderComponent } from './process-instance-header.component';
-import { ProcessTestingModule } from '../../testing/process.testing.module';
+import { ProcessTestingModule } from '../../../testing/process.testing.module';
 
 describe('ProcessInstanceHeaderComponent', () => {
     let component: ProcessInstanceHeaderComponent;
