+++ lib/process-services/src/lib/process-list/components/process-instance-tasks/process-instance-tasks.component.spec.ts
@@ -19,10 +19,10 @@ import { SimpleChange } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { of } from 'rxjs';
-import { taskDetailsMock } from '../../mock';
-import { ProcessService } from './../services/process.service';
+import { taskDetailsMock } from '../../../mock';
+import { ProcessService } from '../../services/process.service';
 import { ProcessInstanceTasksComponent } from './process-instance-tasks.component';
-import { ProcessTestingModule } from '../../testing/process.testing.module';
+import { ProcessTestingModule } from '../../../testing/process.testing.module';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { MatListItemHarness } from '@angular/material/list/testing';
