+++ lib/process-services/src/lib/process-list/components/process-list/process-list.component.spec.ts
@@ -30,9 +30,9 @@ import {
     DataCellEvent,
     ObjectDataColumn
 } from '@alfresco/adf-core';
-import { fakeProcessInstance, fakeProcessInstancesWithNoName, fakeProcessInstancesEmpty, fakeProcessColumnSchema } from '../../mock';
-import { ProcessService } from '../services/process.service';
-import { ProcessTestingModule } from '../../testing/process.testing.module';
+import { fakeProcessInstance, fakeProcessInstancesWithNoName, fakeProcessInstancesEmpty, fakeProcessColumnSchema } from '../../../mock';
+import { ProcessService } from '../../services/process.service';
+import { ProcessTestingModule } from '../../../testing/process.testing.module';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { MatProgressSpinnerHarness } from '@angular/material/progress-spinner/testing';
