+++ lib/process-services/src/lib/process-list/components/start-process/start-process.component.spec.ts
@@ -18,15 +18,15 @@
 import { SimpleChange } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { AppConfigService } from '@alfresco/adf-core';
-import { AppsProcessService } from '../../app-list/services/apps-process.service';
+import { AppsProcessService } from '../../../app-list/services/apps-process.service';
 import { of, throwError } from 'rxjs';
 import { MatSelectChange } from '@angular/material/select';
-import { ProcessService } from '../services/process.service';
-import { newProcess, taskFormMock, testProcessDef, testMultipleProcessDefs, testProcessDefWithForm, testProcessDefinitions } from '../../mock';
+import { ProcessService } from '../../services/process.service';
+import { newProcess, taskFormMock, testProcessDef, testMultipleProcessDefs, testProcessDefWithForm, testProcessDefinitions } from '../../../mock';
 import { StartProcessInstanceComponent } from './start-process.component';
-import { ProcessTestingModule } from '../../testing/process.testing.module';
-import { deployedApps } from '../../mock/apps-list.mock';
-import { ActivitiContentService } from '../../form/services/activiti-alfresco.service';
+import { ProcessTestingModule } from '../../../testing/process.testing.module';
+import { deployedApps } from '../../../mock/apps-list.mock';
+import { ActivitiContentService } from '../../../form/services/activiti-alfresco.service';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { MatFormFieldHarness } from '@angular/material/form-field/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
