+++ lib/process-services/src/lib/process-list/public-api.ts
@@ -16,21 +16,21 @@
  */
 
 // component
-import { ProcessInstanceListComponent } from './components/process-list.component';
-import { ProcessFiltersComponent } from './components/process-filters.component';
-import { ProcessInstanceDetailsComponent } from './components/process-instance-details.component';
-import { ProcessInstanceHeaderComponent } from './components/process-instance-header.component';
-import { ProcessInstanceTasksComponent } from './components/process-instance-tasks.component';
-import { StartProcessInstanceComponent } from './components/start-process.component';
-import { ProcessAuditDirective } from './components/process-audit.directive';
+import { ProcessInstanceListComponent } from './components/process-list/process-list.component';
+import { ProcessFiltersComponent } from './components/process-filters/process-filters.component';
+import { ProcessInstanceDetailsComponent } from './components/process-instance-details/process-instance-details.component';
+import { ProcessInstanceHeaderComponent } from './components/process-instance-header/process-instance-header.component';
+import { ProcessInstanceTasksComponent } from './components/process-instance-tasks/process-instance-tasks.component';
+import { StartProcessInstanceComponent } from './components/start-process/start-process.component';
+import { ProcessAuditDirective } from './components/process-audit/process-audit.directive';
 
-export * from './components/process-filters.component';
-export * from './components/process-instance-details.component';
-export * from './components/process-audit.directive';
-export * from './components/process-instance-header.component';
-export * from './components/process-instance-tasks.component';
-export * from './components/process-list.component';
-export * from './components/start-process.component';
+export * from './components/process-filters/process-filters.component';
+export * from './components/process-instance-details/process-instance-details.component';
+export * from './components/process-audit/process-audit.directive';
+export * from './components/process-instance-header/process-instance-header.component';
+export * from './components/process-instance-tasks/process-instance-tasks.component';
+export * from './components/process-list/process-list.component';
+export * from './components/start-process/start-process.component';
 
 // services
 export * from './services/process.service';
